disease-main/requirements.txt
gtts
google-generativeai
pillow
pandas
streamlit
